+++ src/components/MultiSelect.tsx
@@ -3,9 +3,15 @@ import { Text, TouchableOpacity, View } from 'react-native';
 import { useThemeStyles } from '../styles';
 import type { Data, SelectorRect, MultiSelectProperties } from '../types';
 import SelectionList from './SelectionList';
+import Svg, { Path } from 'react-native-svg';
 
 /* Renders a multi-selector component. Takes in props defined in the MultiSelectProperties type. */
 const MultiSelect = (props: MultiSelectProperties): React.JSX.Element => {
+  const defaultPlaceholderText = 'Click me';
+  const disabledOpacity = .5;
+  const enabledOpacity = 1;
+  const refRectYOffset = 5;
+  const defaultListHeight = 200;
   const style = useThemeStyles(props.theme ?? 'system');
   const ref = useRef<TouchableOpacity>(null);
   const [listDisplay, setListDisplay] = useState<boolean>(false);
@@ -27,9 +33,9 @@ const MultiSelect = (props: MultiSelectProperties): React.JSX.Element => {
     ref.current?.measureInWindow((x, y, width, height) => {
       setRefRect({
         x: x,
-        y: y - 5,
+        y: y - refRectYOffset,
         width: props.boxStyle?.width ?? width,
-        height: height + 10,
+        height: height + 2*refRectYOffset,
       });
       if (display)
         setListDisplay(true);
@@ -42,7 +48,7 @@ const MultiSelect = (props: MultiSelectProperties): React.JSX.Element => {
         style={[
           style.selectorBox,
           props.boxStyle,
-          {opacity: props.disabled ? .5 : 1},
+          {opacity: props.disabled ? disabledOpacity : enabledOpacity},
         ]}
         disabled={props.disabled}
         onPress={() => updatePos(true)}
@@ -72,17 +78,20 @@ const MultiSelect = (props: MultiSelectProperties): React.JSX.Element => {
           : <Text
               style={[style.selectorText, props.boxTextStyle]}
             >
-              {props.placeholderText ?? 'Click me'}
+              {props.placeholderText ?? defaultPlaceholderText}
             </Text>
         }
-        <Text
-          style={{
-            ...style.arrow,
-            color: props.dropdownArrowColor ?? style.arrow.color,
-          }}
-        >
-          {listDisplay ? 'ᨈ' : 'ᨆ'}
-        </Text>
+        <View style={{ position: 'absolute', right: 0, paddingBottom: 4 }}>
+          {listDisplay ? (
+            // This is the up arrow "ᨈ"
+            <Svg width={25} height={25} viewBox="0 0 25 25" fill="none">
+              <Path d="M17 14l-5-5-5 5" stroke={props.dropdownArrowColor ?? style.arrow.color} strokeWidth={2} strokeLinecap="round" strokeLinejoin="round" />
+            </Svg>) : (
+            // This is the down arrow "ᨆ"
+            <Svg width={25} height={25} viewBox="0 0 25 25" fill="none">
+              <Path d="M7 10l5 5 5-5" stroke={props.dropdownArrowColor ?? style.arrow.color} strokeWidth={2} strokeLinecap="round" strokeLinejoin="round" />
+            </Svg>)}
+        </View>
       </TouchableOpacity>
       <SelectionList
         styles={{
@@ -105,7 +114,7 @@ const MultiSelect = (props: MultiSelectProperties): React.JSX.Element => {
           }
           setSelected([]);
         }}
-        listHeight={props.listHeight ?? 200}
+        listHeight={props.listHeight ?? defaultListHeight}
         display={listDisplay}
         searchable={!!props.searchable}
         hide={() => setListDisplay(false)}